+++ src/App.ts
@@ -4,7 +4,6 @@ import { config } from "@golemio/core/dist/output-gateway/config/config";
 import { mongooseConnection, sequelizeConnection } from "@golemio/core/dist/output-gateway/database";
 import { log, requestLogger } from "@golemio/core/dist/output-gateway/Logger";
 import { CacheMiddleware, RedisConnector } from "@golemio/core/dist/output-gateway/redis";
-import { RouterBuilder } from "@golemio/core/dist/output-gateway/routes";
 import express, { NextFunction, Request, Response } from "@golemio/core/dist/shared/express";
 import { CustomError, ErrorHandler, HTTPErrorHandler, ICustomErrorObject } from "@golemio/core/dist/shared/golemio-errors";
 import { createLightship, Lightship } from "@golemio/core/dist/shared/lightship";
@@ -12,7 +11,6 @@ import sentry from "@golemio/core/dist/shared/sentry";
 import compression from "compression";
 import http from "http";
 import swaggerUi from "swagger-ui-express";
-import { generalRoutes } from "./general-routes";
 import {
     airQualityRouter,
     bicycleCountersRouter,
@@ -26,14 +24,13 @@ import {
     municipalLibrariesRouter,
     municipalPoliceStationsRouter,
     parkingsRouter,
-    tskParkingLotsLegacyRouter,
-    parkingZonesRouter,
     pedestriansRouter,
     playgroundsRouter,
     sharedBikesGbfsRouter,
     sharedBikesRouter,
     sharedCarsRouter,
     trafficRouter,
+    tskParkingLotsLegacyRouter,
     wasteCollectionLegacyRouter,
     wasteCollectionYardsRouter,
 } from "./routers";
@@ -221,7 +218,6 @@ export default class App extends BaseApp {
         this.express.use("/medicalinstitutions", medicalInstitutionsRouter);
         this.express.use("/municipalauthorities", municipalAuthoritiesRouter);
         this.express.use("/municipallibraries", municipalLibrariesRouter);
-        this.express.use("/parkingzones", parkingZonesRouter);
         this.express.use("/gardens", gardensRouter);
         this.express.use("/sortedwastestations", wasteCollectionLegacyRouter);
         this.express.use("/wastecollectionyards", wasteCollectionYardsRouter);
@@ -248,11 +244,6 @@ export default class App extends BaseApp {
             swaggerUi.serveFiles(require("../docs/generated/public-openapi.json"), {}),
             swaggerUi.setup(require("../docs/generated/public-openapi.json"))
         );
-
-        // Create general routes through builder
-        const builder: RouterBuilder = new RouterBuilder(defaultRouter);
-        builder.LoadData(generalRoutes);
-        builder.BuildAllRoutes();
     };
 
     /**