+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Changed
 
--   removed modules: bicycle-parkings, parking-zones, public-toilets
+-   removed modules: bicycle-parkings, parking-zones, public-toilets, shared-cars, meteosensors, traffic-cameras
+-   removed general-routes
 
 ## [3.4.7] - 2023-04-03